requirements.txt
aiohappyeyeballs==2.4.3
aiohttp==3.10.8
aiosignal==1.3.1
altair==5.4.1
annotated-types==0.7.0
anthropic==0.34.2
anyio==4.6.0
attrs==24.2.0
blinker==1.8.2
cachetools==5.5.0
certifi==2024.8.30
charset-normalizer==3.3.2
click==8.1.7
dataclasses-json==0.6.7
defusedxml==0.7.1
distro==1.9.0
docstring_parser==0.16
filelock==3.16.1
frozenlist==1.4.1
fsspec==2024.9.0
gitdb==4.0.11
GitPython==3.1.43
google-api-core==2.20.0
google-auth==2.35.0
google-cloud-aiplatform==1.68.0
google-cloud-bigquery==3.26.0
google-cloud-core==2.4.1
google-cloud-resource-manager==1.12.5
google-cloud-storage==2.18.2
google-crc32c==1.6.0
google-resumable-media==2.7.2
googleapis-common-protos==1.65.0
grpc-google-iam-v1==0.13.1
grpcio==1.66.2
grpcio-status==1.66.2
h11==0.14.0
httpcore==1.0.6
httpx==0.27.2
httpx-sse==0.4.0
huggingface-hub==0.25.1
idna==3.10
Jinja2==3.1.4
jiter==0.5.0
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
langchain==0.3.1
langchain-anthropic==0.2.1
langchain-community==0.3.1
langchain-core==0.3.7
langchain-google-vertexai==2.0.2
langchain-openai==0.2.1
langchain-text-splitters==0.3.0
langgraph==0.2.31
langgraph-checkpoint==1.0.14
langsmith==0.1.129
markdown-it-py==3.0.0
MarkupSafe==2.1.5
marshmallow==3.22.0
mdurl==0.1.2
msgpack==1.1.0
multidict==6.1.0
mypy-extensions==1.0.0
narwhals==1.9.0
numpy==1.26.4
openai==1.50.2
orjson==3.10.7
packaging==24.1
pandas==2.2.3
pillow==10.4.0
proto-plus==1.24.0
protobuf==5.28.2
pyarrow==17.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
pydantic==2.9.2
pydantic-settings==2.5.2
pydantic_core==2.23.4
pydeck==0.9.1
Pygments==2.18.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
PyYAML==6.0.2
referencing==0.35.1
regex==2024.9.11
requests==2.32.3
rich==13.9.1
rpds-py==0.20.0
rsa==4.9
shapely==2.0.6
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
SQLAlchemy==2.0.35
streamlit==1.39.0
tavily-python==0.5.0
tenacity==8.5.0
tiktoken==0.7.0
tokenizers==0.20.0
toml==0.10.2
tornado==6.4.1
tqdm==4.66.5
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
yarl==1.13.1
